Addressing review comments, and fixing CI
